requirements.txt
Django==4.1.3
django-import-export==3.0.1
djangorestframework==3.14.0